move some functions non-minio out
